完善可追溯 RAG 工作台与开源交付
完善文档导入、混合检索、重排、拒答和引用校验链路。

补充工作台交互、Docker 启动、评测工具、可靠性测试及开源项目文档。

+++ app/import_process/agent/nodes/node_item_name_recognition.py
@@ -133,7 +133,10 @@ def step_4(item_name, file_title, dense_vector, sparse_vector, kb_id="", documen
     )
     milvus_client.delete(milvus_config.item_name_collection, filter=delete_filter)
     #存储数据
-    normalized_dense_vector = np.asarray(dense_vector, dtype=np.float16)
+    # Milvus FLOAT_VECTOR fields and query paths use float32. Keeping this
+    # boundary explicit avoids GPU/driver-specific Half values leaking into
+    # storage or later search calls.
+    normalized_dense_vector = np.asarray(dense_vector, dtype=np.float32)
 
     data = [
             {
